server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = "==3.11.9"
dependencies = [
    "cloudscraper>=1.2.71",
    "flask>=3.1.0",
    "lxml>=5.3.2",
    "ollama>=0.4.7",
    "openai-whisper>=20240930",
    "pyannote-audio>=3.3.2",
    "pypdf>=5.4.0",
]
